+++ packages/auth/android/src/main/java/io/invertase/firebase/auth/ReactNativeFirebaseAuthModule.java
@@ -21,6 +21,7 @@ import android.app.Activity;
 import android.net.Uri;
 import android.os.Parcel;
 import android.util.Log;
+import androidx.annotation.NonNull;
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
@@ -42,6 +43,7 @@ import com.google.firebase.auth.FacebookAuthProvider;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthException;
 import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException;
+import com.google.firebase.auth.FirebaseAuthMultiFactorException;
 import com.google.firebase.auth.FirebaseAuthProvider;
 import com.google.firebase.auth.FirebaseAuthSettings;
 import com.google.firebase.auth.FirebaseUser;
@@ -49,9 +51,17 @@ import com.google.firebase.auth.FirebaseUserMetadata;
 import com.google.firebase.auth.GetTokenResult;
 import com.google.firebase.auth.GithubAuthProvider;
 import com.google.firebase.auth.GoogleAuthProvider;
+import com.google.firebase.auth.MultiFactorAssertion;
+import com.google.firebase.auth.MultiFactorInfo;
+import com.google.firebase.auth.MultiFactorResolver;
+import com.google.firebase.auth.MultiFactorSession;
 import com.google.firebase.auth.OAuthProvider;
 import com.google.firebase.auth.PhoneAuthCredential;
+import com.google.firebase.auth.PhoneAuthOptions;
 import com.google.firebase.auth.PhoneAuthProvider;
+import com.google.firebase.auth.PhoneMultiFactorAssertion;
+import com.google.firebase.auth.PhoneMultiFactorGenerator;
+import com.google.firebase.auth.PhoneMultiFactorInfo;
 import com.google.firebase.auth.TwitterAuthProvider;
 import com.google.firebase.auth.UserInfo;
 import com.google.firebase.auth.UserProfileChangeRequest;
@@ -59,6 +69,8 @@ import io.invertase.firebase.common.ReactNativeFirebaseEvent;
 import io.invertase.firebase.common.ReactNativeFirebaseEventEmitter;
 import io.invertase.firebase.common.ReactNativeFirebaseModule;
 import io.invertase.firebase.common.SharedUtils;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -72,13 +84,23 @@ import javax.annotation.Nullable;
 
 @SuppressWarnings({"ThrowableResultOfMethodCallIgnored", "JavaDoc"})
 class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
+  // Easier to use would be Instant and DateTimeFormatter, but these are only available in API 26+
+  // According to https://stackoverflow.com/a/2202300 this is not the optimal solution, but we only
+  // get a unix timestamp so we can hardcode the timezone.
+  public static final SimpleDateFormat ISO_8601_FORMATTER =
+      new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ");
+
   private static final String TAG = "Auth";
   private static HashMap<String, FirebaseAuth.AuthStateListener> mAuthListeners = new HashMap<>();
   private static HashMap<String, FirebaseAuth.IdTokenListener> mIdTokenListeners = new HashMap<>();
+  private static HashMap<String, String> emulatorConfigs = new HashMap<>();
   private String mVerificationId;
   private String mLastPhoneNumber;
   private PhoneAuthProvider.ForceResendingToken mForceResendingToken;
   private PhoneAuthCredential mCredential;
+
+  private final HashMap<String, MultiFactorResolver> mCachedResolvers = new HashMap<>();
+  private final HashMap<String, MultiFactorSession> mMultiFactorSessions = new HashMap<>();
 
   ReactNativeFirebaseAuthModule(ReactApplicationContext reactContext) {
     super(reactContext, TAG);
@@ -119,6 +141,9 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
       firebaseAuth.removeIdTokenListener(mAuthListener);
       idTokenListenerIterator.remove();
     }
+
+    mCachedResolvers.clear();
+    mMultiFactorSessions.clear();
   }
 
   /** Add a new auth state listener - if one doesn't exist already */
@@ -929,6 +954,228 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
   }
 
   @ReactMethod
+  public void getSession(final String appName, final Promise promise) {
+    FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
+    FirebaseAuth firebaseAuth = FirebaseAuth.getInstance(firebaseApp);
+    firebaseAuth
+        .getCurrentUser()
+        .getMultiFactor()
+        .getSession()
+        .addOnCompleteListener(
+            task -> {
+              if (!task.isSuccessful()) {
+                rejectPromiseWithExceptionMap(promise, task.getException());
+                return;
+              }
+
+              final MultiFactorSession session = task.getResult();
+              final String sessionId = Integer.toString(session.hashCode());
+              mMultiFactorSessions.put(sessionId, session);
+
+              promise.resolve(sessionId);
+            });
+  }
+
+  @ReactMethod
+  public void verifyPhoneNumberWithMultiFactorInfo(
+      final String appName, final String hintUid, final String sessionKey, final Promise promise) {
+    final MultiFactorResolver resolver = mCachedResolvers.get(sessionKey);
+    if (resolver == null) {
+      // See https://firebase.google.com/docs/reference/node/firebase.auth.multifactorresolver for
+      // the error code
+      rejectPromiseWithCodeAndMessage(
+          promise,
+          "invalid-multi-factor-session",
+          "No resolver for session found. Is the session id correct?");
+      return;
+    }
+
+    MultiFactorInfo selectedHint = null;
+    for (MultiFactorInfo multiFactorInfo : resolver.getHints()) {
+      if (hintUid.equals(multiFactorInfo.getUid())) {
+        selectedHint = multiFactorInfo;
+        break;
+      }
+    }
+
+    if (selectedHint == null) {
+      rejectPromiseWithCodeAndMessage(
+          promise,
+          "multi-factor-info-not-found",
+          "The user does not have a second factor matching the identifier provided.");
+      return;
+    }
+
+    if (!PhoneMultiFactorGenerator.FACTOR_ID.equals(selectedHint.getFactorId())) {
+      rejectPromiseWithCodeAndMessage(
+          promise, "unknown", "Unsupported second factor. Only phone factors are supported.");
+      return;
+    }
+
+    final Activity activity = getCurrentActivity();
+    final PhoneAuthOptions phoneAuthOptions =
+        PhoneAuthOptions.newBuilder()
+            .setActivity(activity)
+            .setMultiFactorHint((PhoneMultiFactorInfo) selectedHint)
+            .setTimeout(30L, TimeUnit.SECONDS)
+            .setMultiFactorSession(resolver.getSession())
+            .setCallbacks(
+                new PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
+                  @Override
+                  public void onCodeSent(
+                      @NonNull String verificationId,
+                      @NonNull PhoneAuthProvider.ForceResendingToken forceResendingToken) {
+                    promise.resolve(verificationId);
+                  }
+
+                  @Override
+                  public void onVerificationCompleted(
+                      @NonNull PhoneAuthCredential phoneAuthCredential) {
+                    resolveMultiFactorCredential(phoneAuthCredential, sessionKey, promise);
+                  }
+
+                  @Override
+                  public void onVerificationFailed(@NonNull FirebaseException e) {
+                    promiseRejectAuthException(promise, e);
+                  }
+                })
+            .build();
+
+    PhoneAuthProvider.verifyPhoneNumber(phoneAuthOptions);
+  }
+
+  @ReactMethod
+  public void verifyPhoneNumberForMultiFactor(
+      final String appName,
+      final String phoneNumber,
+      final String sessionKey,
+      final Promise promise) {
+    final MultiFactorSession multiFactorSession = mMultiFactorSessions.get(sessionKey);
+    if (multiFactorSession == null) {
+      rejectPromiseWithCodeAndMessage(promise, "unknown", "can't find session for provided key");
+      return;
+    }
+
+    final PhoneAuthOptions phoneAuthOptions =
+        PhoneAuthOptions.newBuilder()
+            .setPhoneNumber(phoneNumber)
+            .setActivity(getCurrentActivity())
+            .setTimeout(30L, TimeUnit.SECONDS)
+            .setMultiFactorSession(multiFactorSession)
+            .requireSmsValidation(true)
+            .setCallbacks(
+                new PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
+                  @Override
+                  public void onVerificationCompleted(
+                      @NonNull PhoneAuthCredential phoneAuthCredential) {
+                    // We can't handle this flow in the JS part if we want to be compatible with
+                    // the firebase-js-sdk. If we set the requireSmsValidation option to true
+                    // this code should not be executed.
+                    rejectPromiseWithCodeAndMessage(
+                        promise, "not-implemented", "This is currently not supported.");
+                  }
+
+                  @Override
+                  public void onVerificationFailed(@NonNull FirebaseException e) {
+                    promiseRejectAuthException(promise, e);
+                  }
+
+                  @Override
+                  public void onCodeSent(
+                      @NonNull String verificationId,
+                      @NonNull PhoneAuthProvider.ForceResendingToken forceResendingToken) {
+                    promise.resolve(verificationId);
+                  }
+                })
+            .build();
+
+    PhoneAuthProvider.verifyPhoneNumber(phoneAuthOptions);
+  }
+
+  @ReactMethod
+  public void finalizeMultiFactorEnrollment(
+      final String appName,
+      final String verificationId,
+      final String verificationCode,
+      @Nullable final String displayName,
+      final Promise promise) {
+    FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
+    FirebaseAuth firebaseAuth = FirebaseAuth.getInstance(firebaseApp);
+
+    final PhoneAuthCredential phoneAuthCredential =
+        PhoneAuthProvider.getCredential(verificationId, verificationCode);
+    final PhoneMultiFactorAssertion assertion =
+        PhoneMultiFactorGenerator.getAssertion(phoneAuthCredential);
+    firebaseAuth
+        .getCurrentUser()
+        .getMultiFactor()
+        .enroll(assertion, displayName)
+        .addOnCompleteListener(
+            task -> {
+              if (!task.isSuccessful()) {
+                rejectPromiseWithExceptionMap(promise, task.getException());
+              }
+
+              // Need to reload user to make it all visible?
+              promise.resolve("yes");
+            });
+  }
+  /**
+   * This method is intended to resolve a {@link PhoneAuthCredential} obtained through a
+   * multi-factor authentication flow. A credential can either be obtained using:
+   *
+   * <ul>
+   *   <li>{@link PhoneAuthProvider#getCredential(String, String)}
+   *   <li>or {@link
+   *       com.google.firebase.auth.PhoneAuthProvider.OnVerificationStateChangedCallbacks#onVerificationCompleted(PhoneAuthCredential)}
+   * </ul>
+   *
+   * @param authCredential
+   * @param sessionKey An identifier for the session the flow belongs to. Used to look up the {@link
+   *     MultiFactorResolver}
+   * @param promise
+   */
+  private void resolveMultiFactorCredential(
+      final PhoneAuthCredential authCredential, final String sessionKey, final Promise promise) {
+    final MultiFactorAssertion multiFactorAssertion =
+        PhoneMultiFactorGenerator.getAssertion(authCredential);
+
+    final MultiFactorResolver resolver = mCachedResolvers.get(sessionKey);
+    if (resolver == null) {
+      rejectPromiseWithCodeAndMessage(
+          promise,
+          "invalid-multi-factor-session",
+          "No resolver for session found. Is the session id correct?");
+      return;
+    }
+
+    resolver
+        .resolveSignIn(multiFactorAssertion)
+        .addOnCompleteListener(
+            task -> {
+              if (task.isSuccessful()) {
+                AuthResult authResult = task.getResult();
+                promiseWithAuthResult(authResult, promise);
+              } else {
+                promiseRejectAuthException(promise, task.getException());
+              }
+            });
+  }
+
+  @ReactMethod
+  public void resolveMultiFactorSignIn(
+      final String appName,
+      final String session,
+      final String verificationId,
+      final String verificationCode,
+      final Promise promise) {
+
+    final PhoneAuthCredential credential =
+        PhoneAuthProvider.getCredential(verificationId, verificationCode);
+    resolveMultiFactorCredential(credential, session, promise);
+  }
+
+  @ReactMethod
   public void confirmationResultConfirm(
       String appName, final String verificationCode, final Promise promise) {
     FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
@@ -1561,10 +1808,13 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
 
   @ReactMethod
   public void useEmulator(String appName, String host, int port) {
-    Log.d(TAG, "useEmulator");
-    FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
-    FirebaseAuth firebaseAuth = FirebaseAuth.getInstance(firebaseApp);
-    firebaseAuth.useEmulator(host, port);
+
+    if (emulatorConfigs.get(appName) == null) {
+      emulatorConfigs.put(appName, "true");
+      FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
+      FirebaseAuth firebaseAuth = FirebaseAuth.getInstance(firebaseApp);
+      firebaseAuth.useEmulator(host, port);
+    }
   }
 
   /* ------------------
@@ -1637,6 +1887,7 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
       authResultMap.putMap("user", userMap);
 
       promise.resolve(authResultMap);
+
     } else {
       promiseNoUser(promise, true);
     }
@@ -1650,7 +1901,14 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
    */
   private void promiseRejectAuthException(Promise promise, Exception exception) {
     WritableMap error = getJSError(exception);
-    rejectPromiseWithCodeAndMessage(promise, error.getString("code"), error.getString("message"));
+    final String sessionId = error.getString("sessionId");
+    final MultiFactorResolver multiFactorResolver = mCachedResolvers.get(sessionId);
+    WritableMap resolverAsMap = Arguments.createMap();
+    if (multiFactorResolver != null) {
+      resolverAsMap = resolverToMap(sessionId, multiFactorResolver);
+    }
+    rejectPromiseWithCodeAndMessage(
+        promise, error.getString("code"), error.getString("message"), resolverAsMap);
   }
 
   /**
@@ -1737,6 +1995,20 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
       }
     }
 
+    if (exception instanceof FirebaseAuthMultiFactorException) {
+      final FirebaseAuthMultiFactorException multiFactorException =
+          (FirebaseAuthMultiFactorException) exception;
+      // Make sure the error code conforms to the Web API. See
+      // https://firebase.google.com/docs/auth/web/multi-factor#signing_users_in_with_a_second_factor
+      code = "MULTI_FACTOR_AUTH_REQUIRED";
+      final MultiFactorResolver resolver = multiFactorException.getResolver();
+      final String sessionId = Integer.toString(resolver.getSession().hashCode());
+      mCachedResolvers.put(sessionId, resolver);
+      // Passing around a resolver ReadableMap leads to issues when trying to send the data back by
+      // calling Promise#reject. Building the map just before sending solves that issue.
+      error.putString("sessionId", sessionId);
+    }
+
     if (code.equals("UNKNOWN")) {
       if (exception instanceof FirebaseAuthInvalidCredentialsException) {
         code = "INVALID_EMAIL";
@@ -1747,6 +2019,24 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
         code = "TOO_MANY_REQUESTS";
         message = message;
       }
+    }
+
+    // Some message need to be rewritten to match error messages from the Web SDK
+    switch (code) {
+      case "ERROR_UNVERIFIED_EMAIL":
+        message = "This operation requires a verified email.";
+        break;
+      case "ERROR_UNSUPPORTED_FIRST_FACTOR":
+        message =
+            "Enrolling a second factor or signing in with a multi-factor account requires sign-in"
+                + " with a supported first factor.";
+        break;
+      case "ERROR_INVALID_PHONE_NUMBER":
+        message =
+            "The format of the phone number provided is incorrect. Please enter the phone number in"
+                + " a format that can be parsed into E.164 format. E.164 phone numbers are written"
+                + " in the format [+][country code][subscriber number including area code].";
+        break;
     }
 
     code = code.toLowerCase(Locale.ROOT).replace("error_", "").replace('_', '-');
@@ -1877,7 +2167,40 @@ class ReactNativeFirebaseAuthModule extends ReactNativeFirebaseModule {
     }
     userMap.putMap("metadata", metadataMap);
 
+    final WritableArray enrolledFactors = Arguments.createArray();
+    for (final MultiFactorInfo hint : user.getMultiFactor().getEnrolledFactors()) {
+      enrolledFactors.pushMap(multiFactorInfoToMap(hint));
+    }
+    final WritableMap multiFactorMap = Arguments.createMap();
+    multiFactorMap.putArray("enrolledFactors", enrolledFactors);
+    userMap.putMap("multiFactor", multiFactorMap);
+
     return userMap;
+  }
+
+  @NonNull
+  private WritableMap resolverToMap(final String sessionId, final MultiFactorResolver resolver) {
+    final WritableMap result = Arguments.createMap();
+    final WritableArray hints = Arguments.createArray();
+    for (MultiFactorInfo hint : resolver.getHints()) {
+      hints.pushMap(multiFactorInfoToMap(hint));
+    }
+
+    result.putArray("hints", hints);
+    result.putString("session", sessionId);
+    return result;
+  }
+
+  @NonNull
+  private WritableMap multiFactorInfoToMap(MultiFactorInfo hint) {
+    final WritableMap hintMap = Arguments.createMap();
+    final Date enrollmentTime = new Date(hint.getEnrollmentTimestamp() * 1000);
+    hintMap.putString("displayName", hint.getDisplayName());
+    hintMap.putString("enrollmentTime", ISO_8601_FORMATTER.format(enrollmentTime));
+    hintMap.putString("factorId", hint.getFactorId());
+    hintMap.putString("uid", hint.getUid());
+
+    return hintMap;
   }
 
   private ActionCodeSettings buildActionCodeSettings(ReadableMap actionCodeSettings) {