+++ packages/database/android/src/main/java/io/invertase/firebase/database/UniversalFirebaseDatabaseCommon.java
@@ -46,9 +46,14 @@ public class UniversalFirebaseDatabaseCommon {
     setDatabaseConfig(firebaseDatabase, appName, dbURL);
 
     HashMap emulatorConfig = getEmulatorConfig(appName, dbURL);
-    if (emulatorConfig != null) {
+
+    if (emulatorConfig != null && emulatorConfig.get("configured") == null) {
       firebaseDatabase.useEmulator(
           (String) emulatorConfig.get("host"), (Integer) emulatorConfig.get("port"));
+      // The underlying SDK may only be configured once, but with hot-reloads in the
+      // javascript bundle, javascript cannot hold SDK configuration state. Keep track here
+      // so we only configure once
+      emulatorConfig.put("configured", "true");
     }
 
     return firebaseDatabase;